refactor(env): move CACHE_DIR definition to env module
* Moved `CACHE_DIR` from `config` to `env` for better organization.
* Updated imports across various modules to reflect this change.
