+++ backend/open_webui/routers/functions.py
@@ -18,11 +18,10 @@ from open_webui.utils.plugin import (
     replace_imports,
     get_function_module_from_cache,
 )
-from open_webui.config import CACHE_DIR
 from open_webui.constants import ERROR_MESSAGES
 from fastapi import APIRouter, Depends, HTTPException, Request, status
 from open_webui.utils.auth import get_admin_user, get_verified_user
-from open_webui.env import SRC_LOG_LEVELS
+from open_webui.env import SRC_LOG_LEVELS, CACHE_DIR
 from pydantic import BaseModel, HttpUrl
