+++ backend/open_webui/routers/pipelines.py
@@ -18,8 +18,7 @@ from pydantic import BaseModel
 from starlette.responses import FileResponse
 from typing import Optional
 
-from open_webui.env import SRC_LOG_LEVELS, AIOHTTP_CLIENT_SESSION_SSL
-from open_webui.config import CACHE_DIR
+from open_webui.env import SRC_LOG_LEVELS, AIOHTTP_CLIENT_SESSION_SSL, CACHE_DIR
 from open_webui.constants import ERROR_MESSAGES
