+++ backend/open_webui/routers/tools.py
@@ -22,8 +22,8 @@ from open_webui.utils.auth import get_admin_user, get_verified_user
 from open_webui.utils.access_control import has_access, has_permission
 from open_webui.utils.tools import get_tool_servers
 
-from open_webui.env import SRC_LOG_LEVELS
-from open_webui.config import CACHE_DIR, BYPASS_ADMIN_ACCESS_CONTROL
+from open_webui.env import SRC_LOG_LEVELS, CACHE_DIR
+from open_webui.config import BYPASS_ADMIN_ACCESS_CONTROL
 from open_webui.constants import ERROR_MESSAGES
